refactor: split email template tasks
